[CHORE/#66] 홈 화면 내 섹션 컴포넌트로 분리
